Tidy and document current configs
